FedKDC/requirements.txt
flwr[simulation]>=1.8.0
flwr-datasets>=0.3.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
PyYAML>=6.0
tqdm>=4.66
datasets>=2.18
evaluate>=0.4
transformers>=4.39
